Fix a portability issue in greygoo.
This fixes a problem in which we use an x86 specific linker
option for generating .o files for embedded keys.  To make this
portable, this changes Makefile to look at one of the C object
to determine the appropriate binary format to use and store that
as an ld option into a file.  The file is then passed to ld when
we generated the embedded key objects.